\name{leap.template.redirect}
\alias{leap.template.redirect}
\title{
    URL navigation
}
\description{
    Redirect to URL
}
\usage{
    leap.template.redirect(url)
}

\arguments{
    \item{url}{
        URL to redirect to
    }
}
\details{
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    leap.template.redirect("https://mydomain.com")
}

\keyword{template}
