\name{leap.test.run}
\alias{leap.test.run}
\title{
    Running Leap test
}
\description{
    Runs Leap test.
}
\usage{
    leap.test.run(testId, params = list())
}
\arguments{
    \item{testId}{
        Test ID or name.
    }
    \item{params}{
        Named list of parameters you want to pass to the test logic.
    }
}
\details{
}
\value{
    Named list with the test specific return variables.
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    test <- leap.test.run(1)
}
\keyword{test}
