\name{leap.var.getDynamicInputs}
\alias{leap.var.getDynamicInputs}
\alias{c.getDynamicInputs}
\title{
    Dynamic inputs
}
\description{
    Gets vector of names of dynamic inputs in current context
}
\usage{
    leap.var.getDynamicInputs()
}
\arguments{
}
\details{
}
\value{
    Vector of dynamic input names
}
\references{
}
\author{
    Kooky Whiz
}
\note{
}

\seealso{
}
\examples{
    names <- leap.var.getDynamicInputs()
}

\keyword{variable}
